+++ src/BayesFilters/include/BayesFilters/LinearMeasurementModel.h
@@ -15,9 +15,9 @@ class bfl::LinearMeasurementModel : public MeasurementModel
 public:
     virtual ~LinearMeasurementModel() noexcept { };
 
-    virtual Eigen::MatrixXf getMeasurementMatrix() const = 0;
+    virtual Eigen::MatrixXd getMeasurementMatrix() const = 0;
 
-    virtual std::pair<bool, bfl::Data> predictedMeasure(const Eigen::Ref<const Eigen::MatrixXf>& cur_states) const override;
+    virtual std::pair<bool, bfl::Data> predictedMeasure(const Eigen::Ref<const Eigen::MatrixXd>& cur_states) const override;
 
     virtual std::pair<bool, bfl::Data> innovation(const bfl::Data& predicted_measurements, const bfl::Data& measurements) const override;
 };
