+++ src/BayesFilters/include/BayesFilters/PFCorrectionDecorator.h
@@ -18,7 +18,7 @@ public:
 
     void setMeasurementModel(std::unique_ptr<MeasurementModel> measurement_model) override;
 
-    std::pair<bool, Eigen::VectorXf> getLikelihood() override;
+    std::pair<bool, Eigen::VectorXd> getLikelihood() override;
 
 protected:
     PFCorrectionDecorator(std::unique_ptr<PFCorrection> correction) noexcept;
