+++ src/BayesFilters/src/AdditiveStateModel.cpp
@@ -5,7 +5,7 @@
 using namespace bfl;
 using namespace Eigen;
 
-void AdditiveStateModel::motion(const Eigen::Ref<const Eigen::MatrixXf>& cur_states, Eigen::Ref<Eigen::MatrixXf> mot_states)
+void AdditiveStateModel::motion(const Eigen::Ref<const Eigen::MatrixXd>& cur_states, Eigen::Ref<Eigen::MatrixXd> mot_states)
 {
     propagate(cur_states, mot_states);
     mot_states += getNoiseSample(mot_states.cols());
